Add files via upload
Added small support for GCPhone - while i do not have GCPhone on my test server was able to help someone that wanted support for it and he recieved sms alerts.  While testing i couldn't get the GCPhone to accept the coords but did a small change that might help with that and have no way to test it.  If anyone is running GCPhone and can confirm that the sms will set gps from alert i would appericate it.

SpikE